finishing up shaman shoulder runes
